requirements.txt
python-telegram-bot==22.0
aiohttp
fake-useragent
aiofiles